Add unique Id to the NavigationViewItem
